+++ app/src/main/java/com/valimade/skycast/weather/domain/model/realtime/WeatherValues.kt
@@ -9,7 +9,7 @@ data class WeatherValues(
     val humidity: Int,
     val precipitationProbability: Int,
     val pressureSurfaceLevel: Double,
-    val rainIntensity: Int,
+    val rainIntensity: Double,
     val sleetIntensity: Int,
     val snowIntensity: Int,
     val temperature: Double,
